Organized the repo into a folder and implemented automatic selection of clusters
