+++ subtype_mvcc/consensus_clustering.py
@@ -1,0 +1,66 @@
+import bisect
+import numpy as np
+
+from sklearn import mixture
+from itertools import combinations
+
+class ConsensusCluster:
+    def __init__(self, cluster_method, L, K, H, resample_proportion=0.8):
+        self.resample_proportion_ = resample_proportion
+        self.cluster_ = cluster_method
+        self.L_ = L
+        self.K_ = K
+        self.H_ = H
+        self.Mk = None
+        self.Ak = None
+        self.deltaK = None
+        self.bestK = None
+
+    def _internal_resample(self, data, proportion):
+        ids = np.random.choice(
+            range(data.shape[0]), size=int(data.shape[0] * proportion), replace=False)
+        return ids, data[ids, :]
+
+    def fit(self, data):
+        Mk = np.zeros((self.K_ - self.L_, data.shape[0], data.shape[0]))
+        Is = np.zeros((data.shape[0],) * 2)
+        for k in range(self.L_, self.K_):
+            print(k)
+            i_ = k - self.L_
+            for h in range(self.H_):
+                ids, dt = self._internal_resample(data, self.resample_proportion_)
+                Mh = mixture.GaussianMixture(n_components=k, covariance_type='diag', reg_covar = 1e-5).fit_predict(dt)
+                ids_sorted = np.argsort(Mh)
+                sorted_ = Mh[ids_sorted]
+                for i in range(k):
+                    ia = bisect.bisect_left(sorted_, i)
+                    ib = bisect.bisect_right(sorted_, i)
+                    is_ = ids_sorted[ia:ib]
+                    ids_ = np.array(list(combinations(is_, 2))).T
+                    if ids_.size != 0:
+                        Mk[i_, ids_[0], ids_[1]] += 1
+                ids_2 = np.array(list(combinations(ids, 2))).T
+                Is[ids_2[0], ids_2[1]] += 1
+            Mk[i_] /= Is + 1e-8
+            Mk[i_] += Mk[i_].T
+            Mk[i_, range(data.shape[0]), range(
+                data.shape[0])] = 1
+            Is.fill(0)
+        self.Mk = Mk
+        self.Ak = np.zeros(self.K_ - self.L_)
+        for i, m in enumerate(Mk):
+            hist, bins = np.histogram(m.ravel(), density=True)
+            self.Ak[i] = np.sum(h * (b - a)
+                                for b, a, h in zip(bins[1:], bins[:-1], np.cumsum(hist)))
+        self.deltaK = np.array([(Ab - Aa) / Aa if i > 2 else Aa
+                                for Ab, Aa, i in zip(self.Ak[1:], self.Ak[:-1], range(self.L_, self.K_ - 1))])
+        self.bestK = np.argmax(self.deltaK) + \
+                     self.L_ if self.deltaK.size > 0 else self.L_
+
+    def predict(self):
+        return self.cluster_(n_clusters=self.bestK).fit_predict(
+            1 - self.Mk[self.bestK - self.L_])
+
+    def predict_data(self, data):
+        return self.cluster_(n_clusters=self.bestK).fit_predict(
+            data)
